Add specific instructions on how to install devkit
It's not clear when downloading the devkit, what your next steps should be - just extracting it to a folder (the default) is not enough. You also have to run `devkitvars.bat` to add the location of the devkit to your path.

This change adds instructions on the Download page to help new users out.