Add code signature task
Ideally the installer executable should be signed immediately after the build.
However we use Appveyor to build the files but the generation of a code
signature requires the key on a physical smart card.
This is why the signature task is implemented in this repository instead.

Nevertheless we use gpg to ensure that the files haven't changed since they
were built.