Add the current required key ID for code sign and verify validity
I now have two keys on my signature stick: the current one and the one from the last year.
Unfortunately the signature was computed by making use of the new cert, but the old key, so that the signatures were invalid.

Also add a verify step, to ensure that the signature is valid.